support for react version 0.11.0 React v0.11.0 was recently released so the specified peerDependency in rrouter v0.4.0 is failing:
http://facebook.github.io/react/blog/2014/07/17/react-v0.11.html

The only listed breaking change is to getDefaultProps(), but you only appear to be using that in your PathnameRouting-test.js file- and it seems like the react v0.11.0 change won't break that instance.

I've tested a local app and things seem to be working with this package.json change.

NOTE: I originally updated package.json to v0.5.0, but after your comment in the pull request, I kept it at v0.4.0